Extracted schema related utils into `AvroSchemaUtils`
